test: Pull out PriceGetter fixture to support additional tests using the same fixture
